Modify promote pipeline jobs to improve run-time:
- RDL file check is only run if any ".rdl" file or any RDL template file is modified relative to the 'master' branch
- File list generation and check job are only run if any 'compile.yml' is modified relative to the 'master' branch
- Verilator regression is run by spawning each smoke test as a separate run thread to parallelize the regression

Related work items: #494085

+++ tools/scripts/promote_rdl_check.sh
@@ -14,23 +14,35 @@
 # limitations under the License.
 #
 
+if [[ $# -ne 1 ]]; then
+    echo "Error, requires branch name argument"
+    exit 1
+else
+    merge_dest=$1
+fi
+
 cd $AHA_POC_REPO
 
-# Run the HTML Doc generator script (to update the REG macro header files)
-# and the individual reg generator script but then remove the docs directories
-bash $AHA_POC_REPO/tools/scripts/reg_gen.sh
-bash $AHA_POC_REPO/tools/scripts/reg_doc_gen.sh
-rm -rf $AHA_POC_REPO/src/integration/docs
-rm -rf $AHA_POC_REPO/src/soc_ifc/docs
+rdl_mod_count=$(git diff --merge-base ${merge_dest} --name-status | grep -c '\.rdl$\|tools\/templates\/rdl\|reg_gen.sh\|reg_gen.py\|reg_doc_gen.sh\|reg_doc_gen.py')
+if [[ ${rdl_mod_count} -gt 0 ]]; then
+    # Run the HTML Doc generator script (to update the REG macro header files)
+    # and the individual reg generator script but then remove the docs directories
+    bash $AHA_POC_REPO/tools/scripts/reg_gen.sh
+    bash $AHA_POC_REPO/tools/scripts/reg_doc_gen.sh
+    rm -rf $AHA_POC_REPO/src/integration/docs
+    rm -rf $AHA_POC_REPO/src/soc_ifc/docs
 
-# Check for any file changes
-if [[ $(git status -s --untracked-files=all --ignored=traditional -- $AHA_POC_REPO/src/ | wc -l) -gt 0 ]]; then
-  echo "Regenerating reg RDL outputs produced some file changes:";
-  git status -s --untracked-files=all --ignored=traditional;
-  git diff;
-  echo "*****************************************";
-  echo "Review above changes locally and resubmit pipeline";
-  echo "(Hint: Check $AHA_POC_REPO for the above changes)";
-  echo "*****************************************";
-  exit 1;
+    # Check for any file changes
+    if [[ $(git status -s --untracked-files=all --ignored=traditional -- $AHA_POC_REPO/src/ | wc -l) -gt 0 ]]; then
+      echo "Regenerating reg RDL outputs produced some file changes:";
+      git status -s --untracked-files=all --ignored=traditional;
+      git diff;
+      echo "*****************************************";
+      echo "Review above changes locally and resubmit pipeline";
+      echo "(Hint: Check $AHA_POC_REPO for the above changes)";
+      echo "*****************************************";
+      exit 1;
+    fi
+else
+    echo "skipping RDL check since no RDL files were modified"
 fi
